Finish Day 9 Part 1
In the 'nextValue' method, when I was making the recursive call, I was
adding the result to the last element in the 'differences' array which
was making the final sum for each sequence too large by the last element
in the starting array.

Began Part 2.  Essentially the same recursive function as 'nextValue',
'previousValue' subtracts the result of the recursive call from the
first element in the array.